fix: shadowMask

+++ Assets/BXRenderPipline/Shaders/Libiary/UnityInput.hlsl
@@ -58,7 +58,7 @@ CBUFFER_START(UnityPerDraw)
     real4 unity_LightData;
     real4 unity_LightIndices[2];
 
-    float4 unity_ProbesOcclusion;
+    half4 unity_ProbesOcclusion;
 
     float4 unity_SpecCube0_HDR;
